useless cast to type ‘using ImGuiFileBrowserFlags = uint32_t’ {aka ‘unsigned int’} [-Werror=useless-cast]
In file included from /mnt/d/dev/Field-Map-Editor/src/sfml/gui/gui.hpp:7:
/mnt/d/dev/Field-Map-Editor/src/sfml/gui/batch.hpp:52:95: error: useless cast to type ‘using ImGuiFileBrowserFlags = uint32_t’ {aka ‘unsigned int’} [-Werror=useless-cast]
   52 |      ImGui::FileBrowser                                                  m_directory_browser{ static_cast<ImGuiFileBrowserFlags>(
      |                                                                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   53 |        static_cast<std::uint32_t>(ImGuiFileBrowserFlags_SelectDirectory)
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   54 |        | static_cast<std::uint32_t>(ImGuiFileBrowserFlags_CreateNewDir)) };
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/mnt/d/dev/Field-Map-Editor/src/sfml/gui/gui.hpp:83:95: error: useless cast to type ‘using ImGuiFileBrowserFlags = uint32_t’ {aka ‘unsigned int’} [-Werror=useless-cast]
   83 |      ImGui::FileBrowser                                                  m_save_file_browser{ static_cast<ImGuiFileBrowserFlags>(
      |                                                                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   84 |        static_cast<std::uint32_t>(ImGuiFileBrowserFlags_EnterNewFilename)
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   85 |        | static_cast<std::uint32_t>(ImGuiFileBrowserFlags_CreateNewDir)) };
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/mnt/d/dev/Field-Map-Editor/src/sfml/gui/gui.hpp:87:95: error: useless cast to type ‘using ImGuiFileBrowserFlags = uint32_t’ {aka ‘unsigned int’} [-Werror=useless-cast]
   87 |      ImGui::FileBrowser                                                  m_directory_browser{ static_cast<ImGuiFileBrowserFlags>(
      |                                                                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   88 |        static_cast<std::uint32_t>(ImGuiFileBrowserFlags_SelectDirectory)
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   89 |        | static_cast<std::uint32_t>(ImGuiFileBrowserFlags_CreateNewDir)) };
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/sfml/gui/batch.hpp
@@ -50,8 +50,7 @@ class batch
      FutureOfFutureConsumer<std::vector<std::future<std::future<void>>>> m_future_of_future_consumer = {};
      FutureConsumer<std::vector<std::future<void>>>                      m_future_consumer           = {};
      ImGui::FileBrowser                                                  m_directory_browser{ static_cast<ImGuiFileBrowserFlags>(
-       static_cast<std::uint32_t>(ImGuiFileBrowserFlags_SelectDirectory)
-       | static_cast<std::uint32_t>(ImGuiFileBrowserFlags_CreateNewDir)) };
+       ImGuiFileBrowserFlags_SelectDirectory | ImGuiFileBrowserFlags_CreateNewDir) };
 
      void                                                                combo_input_type(int &imgui_id);
      void                                                                combo_output_type(int &imgui_id);
